requirements.txt
django
djangorestframework
scikit-learn
pandas
joblib